+++ trunk/SSEPlus_Root/SSEPlus/include/emulation/SSEPlus_emulation_SSE2.h
@@ -1710,7 +1710,7 @@ SSP_FORCEINLINE __m128i ssp_rot_epi32_SSE2(__m128i a, __m128i b  )
 
     for( n = 0; n < 4; n++ )
     {
-      if( B.s8[n] < 0 )
+      if( B.s32[n] < 0 )
       {
         unsigned int count = (-B.s32[n]) % 32;
         unsigned int carry_count = (32 - count) % 32;
@@ -1739,7 +1739,7 @@ SSP_FORCEINLINE __m128i ssp_rot_epi64_SSE2(__m128i a, __m128i b  )
 
     for( n = 0; n < 2; n++ )
     {
-      if( B.s8[n] < 0 )
+      if( B.s64[n] < 0 )
       {
         unsigned int count = (unsigned int)((-B.s64[n]) % 64);
         unsigned int carry_count = (64 - count) % 64;
